backend/requirements.txt
Django==5.0.14
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-mptt==0.14.0
django-filter==25.1
drf-nested-routers==0.93.4
Pillow==10.2.0
python-decouple==3.8

# Dependencias para optimización de archivos estáticos
django-compressor==4.4
whitenoise==6.6.0
brotlipy==0.7.0
django-storages==1.14.2
boto3==1.34.72
django-cors-headers==4.3.1
django-cache-url==3.4.4  # Para configuración fácil de caché
django-storages-redux==1.3.3  # Soporte adicional para almacenamiento
requests==2.31.0  # Para verificaciones SRI y peticiones HTTP

# Optimización de imágenes
pillow-avif-plugin==1.4.1
imagekitio==4.0.0
opencv-python-headless==4.8.1.78  # Para recorte inteligente (opcional)
face-recognition==1.3.0  # Para detección facial en imágenes (opcional)
